requirements.txt
gradio==4.44.1
gradio_client==1.3.0
numpy==1.26.1
onnx==1.17.0
onnxruntime
Pillow==9.5.0
python-multipart==0.0.9
